Revert "Revert "Fetching data from firebase""
This reverts commit 8b1ab5b42a10bb13afcc24d0268ed28d7d0161ce.
