+++ css/style.css
@@ -1,7 +1,8 @@
 @import "https://fonts.googleapis.com/css2?family=Raleway:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap";
 * {
   margin: 0;
-  font-family: 'Raleway', sans-serif; }
+  font-family: 'Raleway', sans-serif;
+  font-weight: 500; }
 
 body {
   background-image: linear-gradient(20deg, #1b860c, #26f813);
@@ -28,7 +29,8 @@ ul {
   flex-flow: row wrap;
   justify-content: space-around;
   background-image: linear-gradient(180deg, white, #dddddd);
-  border-radius: 2rem; }
+  border-radius: 2rem;
+  margin-bottom: 2rem; }
   @media (min-width: 375px) and (max-width: 640px) {
     .leaderboard-container {
       width: 90%; } }
@@ -42,5 +44,8 @@ ul {
       display: flex;
       flex-flow: row wrap;
       justify-content: space-evenly; }
+    .leaderboard-container .leaderboard-container__table .table-item__winner {
+      font-weight: 700;
+      font-size: 1.2rem; }
 
 /*# sourceMappingURL=style.css.map */
