+++ scss/style.scss
@@ -12,6 +12,7 @@ $tablet-width: 640px;
 *{
     margin: 0;
     font-family: 'Raleway', sans-serif;
+    font-weight: 500;
 }
 
 body{
@@ -47,6 +48,7 @@ ul{
     justify-content: space-around;
     background-image: linear-gradient(180deg, rgb(255, 255, 255), rgb(221, 221, 221));
     border-radius: 2rem;
+    margin-bottom: 2rem;
     @include mobile-device{
         width: 90%;
     }
@@ -64,6 +66,10 @@ ul{
             display: flex;
             flex-flow: row wrap;
             justify-content: space-evenly;
+        }
+        .table-item__winner{
+            font-weight: 700;
+            font-size: 1.2rem;
         }
     }
 }
